+++ ampere/pkb/linux_benchmarks/llm_benchmark.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2025, Ampere Computing LLC
+# Copyright (c) 2026, Ampere Computing LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,14 +18,11 @@
 This is a set of benchmarks that measures performance of llama
 
 """
-
+import logging
 import csv
 import posixpath
 from dataclasses import dataclass
-import logging
 import os
-import uuid
-import time
 from typing import Any, Dict, List
 import glob
 import six
@@ -38,6 +35,14 @@ from perfkitbenchmarker.virtual_machine import BaseVirtualMachine
 from ampere.pkb.common import download_utils
 from ampere.pkb.linux_packages import docker as docker_package
 from ampere.pkb.linux_packages import llama as ampere_llama
+from ampere.pkb.utils import llm_base_utils
+from ampere.pkb.utils import threads_validation
+try:
+    from ampere.pkb_internal.utils import llm_utils as llm_utils_internal
+except ImportError as err:
+    llm_utils_internal = None
+    logging.info(f"Failed to import llm_utils_internal: {err}")
+
 
 BENCHMARK_NAME = "ampere_llama_benchmark"
 
@@ -52,8 +57,6 @@ ampere_llama_benchmark:
 
 FLAGS = flags.FLAGS
 
-INSTABILITY_THRESHOLD = 1.01
-
 model_names = flags.DEFINE_list(f"{BENCHMARK_NAME}_model_names", [], "")
 
 flags.DEFINE_string(f"{BENCHMARK_NAME}_llama_exe_path", None, "")
@@ -62,11 +65,13 @@ threads_per_process_list = flags.DEFINE_list(
     f"{BENCHMARK_NAME}_threads_per_process", [], ""
 )
 
+flags.DEFINE_integer(f"{BENCHMARK_NAME}_number_of_models", 0, "")
+
 batch_sizes_list = flags.DEFINE_list(f"{BENCHMARK_NAME}_batch_size", [], "")
 
 prompt_sizes_list = flags.DEFINE_list(f"{BENCHMARK_NAME}_prompt_size", [], "")
 
-TOKENS = flags.DEFINE_integer(f"{BENCHMARK_NAME}_output_tokens", 256, "")
+flags.DEFINE_integer(f"{BENCHMARK_NAME}_output_tokens", 256, "")
 
 flags.DEFINE_string(f"{BENCHMARK_NAME}_threads_range", "", "")
 
@@ -76,64 +81,11 @@ flags.DEFINE_float(f"{BENCHMARK_NAME}_timeout", 100.00, "")
 
 flags.DEFINE_bool(f"{BENCHMARK_NAME}_stability", False, "")
 
-
-def parse_threads_range(threads_range: str) -> list[int]:
-    """
-    Parses a thread range string into a list of individual thread indices.
-
-    Args:
-        threads_range (str): A string specifying thread index ranges.
-
-    Returns:
-        list[int]: A list of individual thread indices.
-    """
-
-    logging.info("threads_range: %s", threads_range)
-    threads_range = [s.split("-") for s in threads_range.split(",")]
-    logging.info("threads_range: %s", threads_range)
-    if not all(len(s) == 2 for s in threads_range):
-        raise ValueError(
-            "Format of --threads_range argument must be '{idx}-{idx},{idx}-{idx},...', "
-            "e.g. '88-88' to use just thread idx 88"
-        )
-    designated_threads = []
-    for s in threads_range:
-        s_0, s_1 = int(s[0]), int(s[1])
-        if s_1 < s_0:
-            raise ValueError(
-                f"Range {s_0}-{s_1} is not valid, second value has to be equal to or"
-                "greater than the first value"
-            )
-        designated_threads += list(range(s_0, s_1 + 1))
-    logging.info("designated_threads: %s", designated_threads)
-    return designated_threads
-
-
-def check_threads_validity():
-    """
-    Validates the requested thread count against available threads.
-
-    This function checks if the number of threads specified in the benchmark
-    configuration (through `threads_range` and `threads_per_process`) does not
-    exceed the available threads after parsing the `threads_range` argument.
-    If the requested number of threads exceeds the available ones, a
-    `ValueError` is raised.
-
-    It accesses the global flags to get the `threads_range` and `threads_per_process`
-    values, parses the thread range, and then compares the total available threads
-    with the requested threads per process.
-
-    """
-
-    threads_range = FLAGS[f"{BENCHMARK_NAME}_threads_range"].value
-    threads_per_proc_list = FLAGS[f"{BENCHMARK_NAME}_threads_per_process"].value
-    available_threads_list = parse_threads_range(threads_range)
-    if len(available_threads_list) < max(threads_per_proc_list):
-        raise ValueError(
-            f"Requested number of threads ({max(threads_per_proc_list)})"
-            f"exceeds threads available ({len(available_threads_list)})"
-        )
-
+flags.DEFINE_string(
+    f"{BENCHMARK_NAME}_memory_placement",
+    "none",
+    "memory placement policy - 'local','interleave' or 'none'",
+)
 
 @dataclass
 class LlamaResult:
@@ -156,6 +108,7 @@ class LlamaResult:
     e2e_avg_latency_sec: list[float]
     pptg_throughput_tps: list[float]
     concurrency: list[float]
+    min_tps_per_user_per_process: list[float]
     start: list[str]
     finish: list[str]
 
@@ -167,6 +120,7 @@ class LlamaResult:
         Returns:
         """
         llama_csv_result = _parse_csv(llama_results)
+
         return cls(
             n_proc=llama_csv_result.n_proc,
             n_threads=llama_csv_result.n_threads,
@@ -185,6 +139,7 @@ class LlamaResult:
             e2e_avg_latency_sec=llama_csv_result.e2e_avg_latency_sec,
             pptg_throughput_tps=llama_csv_result.pptg_throughput_tps,
             concurrency=llama_csv_result.concurrency,
+            min_tps_per_user_per_process=llama_csv_result.min_tps_per_user_per_process,
             start=llama_csv_result.start,
             finish=llama_csv_result.finish,
         )
@@ -201,6 +156,24 @@ class LlamaResult:
             metadata_new["prompt_size"] = int(self.prompt_size[count_n_proc])
             metadata_sample = metadata | metadata_new
             samples = [
+                sample.Sample(
+                    "Processes",
+                    self.n_proc[count_n_proc],
+                    "",
+                    metadata_sample,
+                ),
+                sample.Sample(
+                    "Threads",
+                    self.n_threads[count_n_proc],
+                    "",
+                    metadata_sample,
+                ),
+                sample.Sample(
+                    "Batch Size",
+                    self.batch_size[count_n_proc],
+                    "",
+                    metadata_sample,
+                ),
                 sample.Sample(
                     "prompt processing throughput",
                     self.pp_throughput_tps[count_n_proc],
@@ -261,6 +234,12 @@ class LlamaResult:
                     "sec",
                     metadata_sample,
                 ),
+                sample.Sample(
+                    "Concurrency",
+                    int(self.concurrency[count_n_proc]),
+                    "",
+                    metadata_sample,
+                ),
             ]
             all_samples.extend(samples)
         return all_samples
@@ -274,7 +253,7 @@ def _parse_csv(llama_results: str) -> LlamaResult:
     tg_throughput_tps,tg_max_latency_sec,tg_avg_latency_sec,
     tg_max_per_token_latency_sec,tg_avg_per_token_latency_sec,
     e2e_max_latency_sec,e2e_avg_latency_sec,
-    pp+tg_throughput_tps,concurrency,start,finish) tuples.
+    pp+tg_throughput_tps,concurrency,min_tps_per_process,start,finish) tuples.
     """
     n_proc: list[int] = []
     n_threads: list[int] = []
@@ -293,10 +272,19 @@ def _parse_csv(llama_results: str) -> LlamaResult:
     e2e_avg_latency_sec: list[float] = []
     pptg_throughput_tps: list[float] = []
     concurrency: list[float] = []
+    min_tps_per_user_per_process: list[float] = []
     start: list[str] = []
     finish: list[str] = []
     csv_fp = six.StringIO(str(llama_results))
     reader = csv.DictReader(csv_fp)
+    try:
+        if FLAGS.sla_per_process:
+            min_tps_per_user_per_process_col = "min_tps_per_user_per_process"
+        else:
+            min_tps_per_user_per_process_col = "tps_per_user"
+    except AttributeError:
+        min_tps_per_user_per_process_col = "tps_per_user"
+
     if frozenset(reader.fieldnames) != frozenset(
         [
             "n_proc",
@@ -316,6 +304,7 @@ def _parse_csv(llama_results: str) -> LlamaResult:
             "e2e_avg_latency_sec",
             "pp+tg_throughput_tps",
             "concurrency",
+            min_tps_per_user_per_process_col,
             "start",
             "finish",
         ]
@@ -339,6 +328,7 @@ def _parse_csv(llama_results: str) -> LlamaResult:
         e2e_avg_latency_sec.append(row["e2e_avg_latency_sec"])
         pptg_throughput_tps.append(row["pp+tg_throughput_tps"])
         concurrency.append(row["concurrency"])
+        min_tps_per_user_per_process.append(row[min_tps_per_user_per_process_col])
         start.append(row["start"])
         finish.append(row["finish"])
     return LlamaResult(
@@ -359,9 +349,11 @@ def _parse_csv(llama_results: str) -> LlamaResult:
         e2e_avg_latency_sec,
         pptg_throughput_tps,
         concurrency,
+        min_tps_per_user_per_process,
         start,
         finish,
     )
+
 
 class LlamaProcessLogResults:
     """Class that represents llama results."""
@@ -379,32 +371,32 @@ class LlamaProcessLogResults:
     e2e_latency: list[float] = []
     row_index: list[int] = []
 
-    def __init__(
-        self,
-        llama_process_logs_file
-        ):
+    def __init__(self, llama_process_logs_file):
         """
         initializes LlamaProcessLogResults object
         """
         try:
-            df = pd.read_csv(llama_process_logs_file, delimiter='|',on_bad_lines='skip')
+            df = pd.read_csv(
+                llama_process_logs_file, delimiter="|", on_bad_lines="skip"
+            )
             for index, row in df.iterrows():
-                if row['Process_number'] != 'Process_number':
-                    self.proc_no.append(int(row['Process_number']))
-                    self.threads_no.append(int(row['threads_per_process']))
-                    self.batch_size.append(int(row['batch_size']))
-                    self.prompt_size.append(int(row['prompt_tokens_per_batch']))
-                    self.output_tokens.append(int(row['tokens_generated_per_batch']))
-                    self.KV_cache_size.append(int(row['KV_cache_size']))
-                    self.time_to_first_token.append(float(row['time_to_first_token']))
-                    self.prompt_processing_throughput.append(float(row['prompt_processing_throughput']))
-                    self.token_gen_latency.append(float(row['token_gen_latency']))
-                    self.token_gen_throughput.append(float(row['token_gen_throughput']))
-                    self.e2e_latency.append(float(row['total_time']))
+                if row["Process_number"] != "Process_number":
+                    self.proc_no.append(int(row["Process_number"]))
+                    self.threads_no.append(int(row["threads_per_process"]))
+                    self.batch_size.append(int(row["batch_size"]))
+                    self.prompt_size.append(int(row["prompt_tokens_per_batch"]))
+                    self.output_tokens.append(int(row["tokens_generated_per_batch"]))
+                    self.KV_cache_size.append(int(row["KV_cache_size"]))
+                    self.time_to_first_token.append(float(row["time_to_first_token"]))
+                    self.prompt_processing_throughput.append(
+                        float(row["prompt_processing_throughput"])
+                    )
+                    self.token_gen_latency.append(float(row["token_gen_latency"]))
+                    self.token_gen_throughput.append(float(row["token_gen_throughput"]))
+                    self.e2e_latency.append(float(row["total_time"]))
                     self.row_index.append(index)
         except TypeError as e:
-            print(e)        
-
+            print(e)
 
     def get_samples(self, metadata: Dict[str, Any]) -> List[sample.Sample]:
         """Return this result as a list of samples."""
@@ -449,10 +441,11 @@ class LlamaProcessLogResults:
                     self.e2e_latency[count_row],
                     "sec",
                     metadata_sample,
-                ),                
+                ),
             ]
             all_samples.extend(samples)
         return all_samples
+
 
 def GetConfig(user_config):
     """Load and return benchmark config.
@@ -472,10 +465,11 @@ def Prepare(benchmark_spec):
     benchmark_spec: The benchmark specification. Contains all data that is
         required to run the benchmark.
     """
-
     servers = benchmark_spec.vm_groups["servers"]
     server = servers[0]
-    check_threads_validity()
+    threads_validation.check_threads_validity(server, BENCHMARK_NAME)
+    if llm_utils_internal:
+        llm_utils_internal.validate_exclusive_run_modes()
     docker_package.Install(server)
 
     # Case 1: Build image
@@ -522,11 +516,8 @@ def Prepare(benchmark_spec):
         )
     docker_package.run_docker(server)
 
-    cmd_exec_installs = ("apt-get update -y &&"
-                         " apt-get install -y numactl")
-    FLAGS[f"{docker_package.PACKAGE_NAME}_shell_type"].value = (
-            "bash"
-            )
+    cmd_exec_installs = "apt-get update -y &&" " apt-get install -y numactl"
+    FLAGS[f"{docker_package.PACKAGE_NAME}_shell_type"].value = "bash"
     FLAGS[f"{docker_package.PACKAGE_NAME}_exec_command"].value = cmd_exec_installs
     docker_package.exec_docker(server)
 
@@ -546,86 +537,116 @@ def Run(benchmark_spec):
     Returns:
         list: A list of benchmark samples or results extracted from the output files.
     """
-
     server = benchmark_spec.vm_groups["servers"][0]
-    benchmark_metadata = {}
-    _run(server)
-    out_dir = posixpath.join(download_utils.INSTALL_DIR, "out_dir")
-    server.RemoteCopy(vm_util.GetTempDir(), out_dir, False)    
-    return _parse_output_files(benchmark_metadata)
+    model = model_names.value[0]
+    sorted_prompt_sizes_list = sorted(prompt_sizes_list.value)
+
+    all_samples = None
+    if llm_utils_internal:
+        all_samples = llm_utils_internal.controller(
+            model, sorted_prompt_sizes_list, server
+        )
+    if not all_samples:
+        _run(server, model)
+        out_dir = posixpath.join(download_utils.INSTALL_DIR, "out_dir")
+        server.RemoteCopy(vm_util.GetTempDir(), out_dir, False)
+        all_samples = _parse_output_files()
+    return all_samples
 
 
-def _run(vm):
+def _run(vm: BaseVirtualMachine, model):
     """
     Executes the Run stage
     """
+    results_llama = []
     threads_range = FLAGS[f"{BENCHMARK_NAME}_threads_range"].value
-    num_available_threads = len(parse_threads_range(threads_range))
-    llama_exe_path = FLAGS[f"{BENCHMARK_NAME}_llama_exe_path"].value
-    stability = FLAGS[f"{BENCHMARK_NAME}_stability"].value
-    volumes = FLAGS[f"{docker_package.PACKAGE_NAME}_volume_names"].value
-    volume_mountpoints = FLAGS[
-        f"{docker_package.PACKAGE_NAME}_volume_mountpoints"
-    ].value
-    output_dir = volumes[2]
-    docker_out_dir = volume_mountpoints[2]
+    threads_range_list = threads_validation.parse_threads_range(vm.cpu_arch, threads_range)
+    available_threads = " ".join(map(str, threads_range_list))
+    for prompt_size in sorted(prompt_sizes_list.value):
+        for batch_size in sorted(batch_sizes_list.value):
+            for num_threads in sorted(threads_per_process_list.value):
+                num_available_threads = len(threads_range_list)
+                num_processes = int(int(num_available_threads) / int(num_threads))
+                if FLAGS[f"{BENCHMARK_NAME}_number_of_models"].value > 0:
+                    num_processes = FLAGS[f"{BENCHMARK_NAME}_number_of_models"].value
+                expt_dict = {
+                        'benchmark': 'ampere_llama_benchmark',
+                        'num_threads': num_threads,
+                        'num_processes': num_processes,
+                        'batch_size': batch_size,
+                        'model': model,
+                        'prompt_size': prompt_size,
+                        'vm': vm,
+                        'available_threads': available_threads,
+                        'tps_per_user':0.0,
+                        }
+                LlamaBase = llm_base_utils.LlamaExperiment(expt_dict)
+                results_llama.extend(
+                    LlamaBase.llama_batched_bench_base_run(
+                        num_processes, num_threads, batch_size, model, prompt_size, vm
+                    ).results
+                )
+    return results_llama
+
+
+def update_llama_metadata(meta_data: dict):
     if FLAGS[f"{docker_package.PACKAGE_NAME}_build_docker_dir"].value:
-        LD_LIBRARY_PATH = ""
+        docker_version = (
+            FLAGS[f"{docker_package.PACKAGE_NAME}_build_docker_image"].value
+            + "-"
+            + FLAGS[f"{docker_package.PACKAGE_NAME}_build_docker_image_version"].value
+        )
     else:
-        LD_LIBRARY_PATH = f"export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:{llama_exe_path} &&"
-    args = {}
-    for model in model_names.value:
-        for prompt_size in sorted(prompt_sizes_list.value):
-            for batch_size in sorted(batch_sizes_list.value):
-                for num_threads in sorted(threads_per_process_list.value):
-                    num_processes = int(num_available_threads / num_threads)
-                    current_case = f"{num_processes} x {num_threads} "
-                    current_case += f"[proc x threads per proc], bs = {batch_size}"
-                    logging.info("\nRunning %s", current_case)
-                    args = {
-                        "model": model,
-                        "prompt_size": prompt_size,
-                        "tokens": TOKENS.value,
-                        "batch_size": batch_size,
-                        "num_processes": num_processes,
-                        "num_threads": num_threads,
-                        "stability": stability,
-                    }
-                    results = Results(vm, args)
-                    while not results.is_stable():
-                        docker_logs_dir = os.path.join(
-                            docker_out_dir, str(uuid.uuid4())
-                        )
-                        logs_dir = os.path.join(
-                            output_dir, docker_logs_dir.split("/")[-1]
-                        )
-                        cmd = (
-                            f"cd / && "
-                            f" {LD_LIBRARY_PATH}"
-                            f" python3 utils/benchmark.py --exe_path {llama_exe_path}/llama-batched-bench "
-                            f" --output_dir {docker_logs_dir} -m models/{model}"
-                            f" -n {str(num_processes)} "
-                            f"-t {str(num_threads)} -b {str(batch_size)} -p {str(prompt_size)}"
-                            f" -k {str(TOKENS.value)} -r {threads_range}"
-                        )
-                        if FLAGS[f"{BENCHMARK_NAME}_stability"].value:
-                            cmd += " --stability"
-                        if FLAGS[f"{BENCHMARK_NAME}_flash_attention"].value:
-                            cmd += " -fa 1"
-                        FLAGS[f"{docker_package.PACKAGE_NAME}_shell_type"].value = (
-                            "bash"
-                        )
-                        FLAGS[f"{docker_package.PACKAGE_NAME}_exec_command"].value = cmd
-                        start = time.time()
-                        docker_package.exec_docker(vm)
-                        finish = time.time()
-                        results.summarize(
-                            vm, logs_dir, start, finish, vm_util.GetTempDir()
-                        )
-                    results.save_csv(vm_util.GetTempDir())
+        docker_version = (
+            FLAGS[f"{docker_package.PACKAGE_NAME}_image"].value
+            + "-"
+            + FLAGS[f"{docker_package.PACKAGE_NAME}_image_version"].value
+        )
+    meta_data["gpu"] = FLAGS[f"{docker_package.PACKAGE_NAME}_gpus"].value
+    meta_data["docker_version"] = docker_version
+    meta_data["gpu_device_id"] = FLAGS[
+            f"{docker_package.PACKAGE_NAME}_gpu_device_id"].value
+    return meta_data
 
 
-def _parse_output_files(benchmark_metadata):
+def collect_all_results_samples(all_files_samples):
+    """function to collect all results in sample for json output"""
+    csv_files = glob.glob(vm_util.GetTempDir() + "/*.csv")
+    if csv_files:
+        for filename in csv_files:
+            csv_file = os.path.basename(filename)
+            model_name = csv_file.split("@")
+            metadata = {
+                "model": model_name[0],
+            }
+            metadata = update_llama_metadata(metadata)
+            csv_file_llama = posixpath.join(vm_util.GetTempDir(), csv_file)
+            with open(csv_file_llama, "r", encoding="utf-8") as output:
+                llama_output_data = output.read()
+            results = LlamaResult.parse_llama_results(llama_output_data)
+            all_files_samples.extend(results.get_samples(metadata))
+    return all_files_samples
+
+
+def collect_process_log_samples(all_files_samples):
+    """function to collect all process wise logs in sample for json output"""
+    process_log_files = glob.glob(vm_util.GetTempDir() + "/*.log")
+    if process_log_files:
+        for log_filename in process_log_files:
+            process_log_file = os.path.basename(log_filename)
+            if "@" in process_log_file:
+                model_name = process_log_file.split("@")
+                metadata = {
+                    "model": model_name[0],
+                }
+                metadata = update_llama_metadata(metadata)
+                log_file_llama = posixpath.join(vm_util.GetTempDir(), process_log_file)
+                log_results = LlamaProcessLogResults(log_file_llama)
+                all_files_samples.extend(log_results.get_samples(metadata))
+    return all_files_samples
+
+
+def _parse_output_files():
     """
     Parses CSV output files from the temporary directory and extracts benchmark samples.
 
@@ -635,254 +656,14 @@ def _parse_output_files(benchmark_metadata):
     results and extracts samples, augmenting the provided benchmark metadata with the
     model name.
 
-    Args:
-        benchmark_metadata (dict): A dictionary of existing metadata to be merged with
-                                   model-specific metadata extracted from each file.
-
     Returns:
         list: A list of sample dictionaries extracted from all parsed CSV files.
     """
 
     all_files_samples = []
-    metadata = {}
-    if FLAGS[f"{docker_package.PACKAGE_NAME}_build_docker_dir"].value:
-        docker_version = FLAGS[f"{docker_package.PACKAGE_NAME}_build_docker_image"].value + "-" + FLAGS[f"{docker_package.PACKAGE_NAME}_build_docker_image_version"].value 
-    else:
-        docker_version = FLAGS[f"{docker_package.PACKAGE_NAME}_image"].value + "-" + FLAGS[f"{docker_package.PACKAGE_NAME}_image_version"].value
-    csv_files = glob.glob(vm_util.GetTempDir() + "/*.csv")
-    if csv_files:
-        for filename in csv_files:
-            csv_file = os.path.basename(filename)
-            model_name = csv_file.split("@")
-            metadata = {
-                "model": model_name[0],
-                "docker_version": docker_version,
-            }
-            csv_file_llama = posixpath.join(vm_util.GetTempDir(), csv_file)
-            with open(csv_file_llama, "r", encoding="utf-8") as output:
-                llama_output_data = output.read()
-            results = LlamaResult.parse_llama_results(llama_output_data)
-            benchmark_metadata = benchmark_metadata | metadata
-            all_files_samples.extend(results.get_samples(benchmark_metadata))
-    process_log_files = glob.glob(vm_util.GetTempDir() + "/*.log")
-    if process_log_files:
-        for log_filename in process_log_files:
-            process_log_file = os.path.basename(log_filename)
-            if "@" in process_log_file:
-                model_name = process_log_file.split("@")
-                metadata = {
-                        "model": model_name[0],
-                        "docker_version": docker_version,
-                        }
-                log_file_llama = posixpath.join(vm_util.GetTempDir(), process_log_file)
-                log_results = LlamaProcessLogResults(log_file_llama)
-                benchmark_metadata = benchmark_metadata | metadata
-                all_files_samples.extend(log_results.get_samples(benchmark_metadata))
+    all_files_samples = collect_all_results_samples(all_files_samples)
+    all_files_samples = collect_process_log_samples(all_files_samples)
     return all_files_samples
-
-
-class Results:
-    """
-    A container for storing and managing benchmark results and metadata.
-
-    Attributes:
-        vm (BaseVirtualMachine): The virtual machine used for the benchmark run.
-        params (dict): A dictionary of benchmark parameters (e.g., batch size, prompt size).
-        tg_runs (list): A list to store token generation run measurements.
-        lines (list): A list to store output lines or logs related to the benchmark.
-        results (list): A list to accumulate parsed or final benchmark results.
-    """
-
-    def __init__(self, vm: BaseVirtualMachine, params):
-        self.params = params
-        self.vm = vm
-        self.tg_runs = []
-        self.lines = []
-        self.results = []
-
-    def summarize(self, vm: BaseVirtualMachine, logs_dir, start, finish, save_logs_dir):
-        """
-        Summarizes the logs by computing averages and sums within a specified range.
-
-        This method processes the log files from the provided directory, using the
-        `start` and `finish` parameters to define the range of interest. It then calculates
-        averages and sums based on the log data, which can be used for further analysis.
-        The processed logs can optionally be saved to a specified directory.
-
-        Args:
-            vm (BaseVirtualMachine): The virtual machine object, potentially used for
-                                      remote log access or command execution.
-            logs_dir (str): The directory containing the log files to be summarized.
-            start (int): The starting index or timestamp to filter logs from.
-            finish (int): The ending index or timestamp to filter logs up to.
-            save_logs_dir (str): The directory to save the processed log summaries (if applicable).
-
-        Returns:
-            None:
-        """
-        time_to_first_token_list = []
-        token_generation_latency_list = []
-        e2e_latency_list = []
-        self.lines = []
-        for n in range(self.params["num_processes"]):
-            line, _ = vm.RemoteCommand(f"head -6 {logs_dir}/log_{n} | tail -1")
-            results = line.strip()[:-1].split("|")
-            line = str(n) + "|" + str(self.params["num_threads"]) + line.strip()[:-1] + "\n"
-            self.lines.append(line)
-            prompt_size = int(results[1])
-            assert prompt_size == self.params["prompt_size"]
-            tokens_generated = int(results[2])
-            assert tokens_generated == TOKENS.value
-            batch_size = int(results[3])
-            assert batch_size == self.params["batch_size"]
-            time_to_first_token_list.append(float(results[5]))
-            token_generation_latency_list.append(float(results[7]))
-            e2e_latency_list.append(float(results[9]))
-        pp_throughput = sum(
-            self.params["batch_size"] * self.params["prompt_size"] / time_to_first_token
-            for time_to_first_token in time_to_first_token_list
-        )
-        avg_pp_latency = sum(time_to_first_token_list) / len(time_to_first_token_list)
-        max_pp_latency = max(time_to_first_token_list)
-        tg_throughput = sum(
-            self.params["batch_size"] * TOKENS.value / lat
-            for lat in token_generation_latency_list
-        )
-        avg_tg_latency = sum(token_generation_latency_list) / len(token_generation_latency_list)
-        max_tg_latency = max(token_generation_latency_list)
-        tg_per_token_lats = [
-            lat / TOKENS.value for lat in token_generation_latency_list
-        ]
-        avg_tg_per_token_latency = sum(tg_per_token_lats) / len(tg_per_token_lats)
-        max_tg_per_token_latency = max(tg_per_token_lats)
-        avg_total_speed = (
-            self.params["num_processes"]
-            * self.params["batch_size"]
-            * (self.params["prompt_size"] + TOKENS.value)
-            / max(
-                time_to_first_token + token_generation_lat
-                for time_to_first_token, token_generation_lat in zip(
-                    time_to_first_token_list, token_generation_latency_list
-                )
-            )
-        )
-        avg_e2e_latency = sum(e2e_latency_list) / len(e2e_latency_list)
-        max_e2e_latency = max(e2e_latency_list)
-
-        self.tg_runs.append(tg_throughput)
-        self.results.append(
-            [
-                self.params["num_processes"],
-                self.params["num_threads"],
-                self.params["batch_size"],
-                self.params["prompt_size"],
-                TOKENS.value,
-                pp_throughput,
-                max_pp_latency,
-                avg_pp_latency,
-                tg_throughput,
-                max_tg_latency,
-                avg_tg_latency,
-                max_tg_per_token_latency,
-                avg_tg_per_token_latency,
-                max_e2e_latency,
-                avg_e2e_latency,
-                avg_total_speed,
-                self.params["batch_size"] * self.params["num_processes"],
-                start,
-                finish,
-            ]
-        )
-        log_filename = (
-            f"{save_logs_dir}/{self.params['model'].split('/')[-1]}@"
-            f"PP{str(self.params['prompt_size'])}@"
-            f"TG{str(TOKENS.value)}@{len(self.tg_runs)}.log"
-        )
-        with open(log_filename, "a", encoding="utf-8") as f1:
-            f1.writelines(
-                    [
-                        "Process_number|",
-                        "threads_per_process|",
-                        "prompt_tokens_per_batch|",
-                        "tokens_generated_per_batch|",
-                        "batch_size|",
-                        "KV_cache_size|",
-                        "time_to_first_token|",
-                        "prompt_processing_throughput|",
-                        "token_gen_latency|",
-                        "token_gen_throughput|",
-                        "total_time|",
-                        "total_speed\n",
-                    ]
-                    )
-            f1.writelines(self.lines)
-        logging.info("Logs saved in %s", log_filename)
-
-    def calc_avg_tg(self, n):
-        """Calculate the average of the first n throughput runs."""
-        return sum(self.tg_runs[:n]) / n
-
-    def is_stable(self):
-        """
-        checks stability of run
-        """
-        logging.info(self.params)
-        runs_completed = len(self.tg_runs)
-        if self.params["stability"] is False and runs_completed > 0:
-            return True
-        if runs_completed < 3:
-            return False
-        prev_avg_tg = self.calc_avg_tg(runs_completed - 1)
-        avg_tg = self.calc_avg_tg(runs_completed)
-        return max(prev_avg_tg / avg_tg, avg_tg / prev_avg_tg) <= INSTABILITY_THRESHOLD
-
-    def save_csv(self, save_dir):
-        """
-        saves csv in save_dir
-        """
-        results_filename = (
-            f"{save_dir}/{self.params['model'].split('/')[-1]}@"
-            f"PP{str(self.params['prompt_size'])}@"
-            f"TG{str(TOKENS.value)}.csv"
-        )
-
-        if os.path.exists(results_filename):
-            first_write = False
-        else:
-            first_write = True
-        with open(results_filename, "a", encoding="utf-8") as f:
-            writer = csv.writer(f)
-            if first_write:
-                writer.writerow(
-                    [
-                        "n_proc",
-                        "n_threads",
-                        "batch_size",
-                        "prompt_size",
-                        "output_tokens",
-                        "pp_throughput_tps",
-                        "pp_max_latency_sec",
-                        "pp_avg_latency_sec",
-                        "tg_throughput_tps",
-                        "tg_max_latency_sec",
-                        "tg_avg_latency_sec",
-                        "tg_max_per_token_latency_sec",
-                        "tg_avg_per_token_latency_sec",
-                        "e2e_max_latency_sec",
-                        "e2e_avg_latency_sec",
-                        "pp+tg_throughput_tps",
-                        "concurrency",
-                        "start",
-                        "finish",
-                    ]
-                )
-            if self.params["stability"] is True:
-                avg_tg = sum(self.tg_runs) / len(self.tg_runs)
-                tg_diff = [abs(avg_tg - tg) for tg in self.tg_runs]
-                writer.writerow(self.results[tg_diff.index(min(tg_diff))])
-            else:
-                writer.writerow(self.results[0])
-        logging.info("Result saved in %s", results_filename)
 
 
 def Cleanup(benchmark_spec):