+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 2026-03-27
+
+### Added
+- `llm_benchmark`: refactor thread range and parsing logic to `llm_base_utils.py`
+- `setup.sh`: upgrade to support Python versions >=3.11.x, update README instructions
+
+### Fixed
+- `requirements.txt`: include `pytz` and lock `setuptools` to keep `pkg_resources` functional in Python >=3.12
+
 ## [2.3.0] - 2025-12-03
 
 ### Added